Rethink universal quantification slightly
In WTy2, being able to universally quantify over types is actually not
that useful. If the quantified type only appears once, it can be
replaced with an existential quantifier, and if the type appears
multiple times, because of how types in WTy2 work, this forces all
terms with that type to have exactly the same value at runtime.

Quantifying over constraints makes a lot more sense.