Minor fixes
Removed the `[n: Num] ...` example. First, correct syntax would be
`n => Num`, but I'm still not sure about this. If `=>` means "instance
head of" - wouldn't this make heterogeneous collections awkward to work
with? I'm not sure `=>` constraints should even be allowed (or `=>`
should be revised to mean simply `implies`).

I think a section of variance is warranted next!