Add section on records
I think I finally have a good set of rules for these. I still find it
a bit awkward that we don't have `(x: Int, y: Int) <: (y: Int, x: Int)`
but with the `?` proposed sugar we can remove the inconvenience of this
AND get nice partial application in one.