Huge spec update, including removing `(<=:)`
In the language as of now, it is already a breaking change to open/
close a type. Therefore, the benefit of having a distinct concept from
subtype and type membership to allow for receiver code to look similar
whether
`type Animal = Cat | Dog;`
or
```
type Animal {};

instance Animal for Cat {};
instance Animal for Dog {};
```
is quite limited. Especially so given `impl(t) = [u: t] u` is so easy
to define.
I still find it slightly conceptually messy that if `Animal` is open
and we have `rufus: Dog`, we do not have `rufus :: Animal` (only
`rufus: impl(Animal)` but with mathematical abstractions. the opposite
actually becomes unnatural (`3` is not a `Monoid`, `(Int, (+))` is!)

Various other tweaks, including describing how operators are curried,
typo fixes, a new page on modules, clarifying rules around importing
`proof`s etc...